Edited README file name and added modifications
Changed the file name from README to README_EDITS. This was done as the web site code is based off of the HTML5 UP! Strata template. Of course, my own personal edits to the code had to be made for customization. All those edits, and future edits, will be added here to the README_EDITS file.